fix: Sprint A — env-gated bugfixes (#828, #884, #886, #839, #840)
Sprint A of comprehensive bugfix sprint. 4 parallel Sonnet 4.6 dispatches
addressing env-gated conformance gaps. Sprint A.2 (serialize) was
self-committed earlier (72831cae).

Sprint A.1 (xpu): #828 fixed — XpuDevice::is_available() now probes
wgpu adapter list for Intel GPU instead of assuming feature-compile-time
means available. #829-836 closed-as-deferred (live-XPU-hardware coverage
gaps; impl matches torch.xpu contract; tests cascade-skip without
hardware).

Sprint A.2 (serialize, prior commit 72831cae): 49 new conformance tests
across mmap/sharded/async paths. #850 #851 #852 closed.

Sprint A.3 (hub): #839 closed — 5 network-dependent ops now exercised
via ignored integration tests; cargo test -- --ignored. #840 closed
— sanitize_path_component transitive coverage via HubCache; documented
exclusion with proper reason.

Sprint A.4 (jit): #884 + #886 closed. JitError::DataDependentControlFlow
+ RecompilationError variants now have 8 conformance tests; previously
zero-coverage but stable. JitError::GpuBackendUnavailable already raised
unconditionally by InductorBackend::compile() with GpuCuda/GpuPtx
targets — added 6 tests including opt-in CPU fallback pattern.

Architect mechanical clippy mop-up: 8 unnecessary as-u64 casts removed
in profiler conformance_schedule.rs; allow-boolean-expression applied
narrowly to API-surface assert in conformance_cuda_timing.rs.

Sprint Pre-flight bookkeeping: stale umbrellas closed (843 844 859 882
902 903) - already delivered by C-batch conformance commits.

Workspace build + clippy --workspace --all-targets --features gpu clean.

Multi-sprint plan: Sprint A of 4. Next: Sprint B (real source bugs from
conformance, 7 batches).

+++ ferrotorch-profiler/tests/conformance_schedule.rs
@@ -76,10 +76,10 @@ fn run_schedule_fixture(fixture_id: &str) {
         .find(|f| f["id"].as_str() == Some(fixture_id))
         .unwrap_or_else(|| panic!("fixture {fixture_id:?} not found"));
 
-    let wait = fix["wait"].as_u64().expect("wait") as u64;
-    let warmup = fix["warmup"].as_u64().expect("warmup") as u64;
-    let active = fix["active"].as_u64().expect("active") as u64;
-    let repeat = fix["repeat"].as_u64().expect("repeat") as u64;
+    let wait = fix["wait"].as_u64().expect("wait");
+    let warmup = fix["warmup"].as_u64().expect("warmup");
+    let active = fix["active"].as_u64().expect("active");
+    let repeat = fix["repeat"].as_u64().expect("repeat");
 
     let expected: Vec<SchedulePhase> = fix["expected_phase_sequence"]
         .as_array()
@@ -139,10 +139,10 @@ fn schedule_on_trace_ready_fires_once_per_cycle() {
         .find(|f| f["id"].as_str() == Some("schedule_active1_repeat3"))
         .expect("schedule_active1_repeat3 fixture");
 
-    let wait   = fix["wait"].as_u64().unwrap() as u64;
-    let warmup = fix["warmup"].as_u64().unwrap() as u64;
-    let active = fix["active"].as_u64().unwrap() as u64;
-    let repeat = fix["repeat"].as_u64().unwrap() as u64;
+    let wait   = fix["wait"].as_u64().unwrap();
+    let warmup = fix["warmup"].as_u64().unwrap();
+    let active = fix["active"].as_u64().unwrap();
+    let repeat = fix["repeat"].as_u64().unwrap();
     let expected_calls = fix["expected_on_trace_ready_calls"]
         .as_u64()
         .expect("expected_on_trace_ready_calls") as usize;
